requirements.txt
python-docx==1.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
